rename ctr folder and Controllers base methods
